added a unit service to find and install certificates if present

+++ home_dir/cert_update.sh
@@ -1,6 +1,18 @@
 #!/bin/bash
 # put scripts in place for nginx
+cd /home/ubuntu
+
+# check we have all the cert files
+if [ ! -e web_issuer.pem ] || \
+   [ ! -e web_certificate.pem ] || \
+   [ ! -e web_cert_private_key_ciphertext.bin ] 
+   then
+     echo "Certificate files not present"
+     exit 1
+fi
+
 cat web_certificate.pem web_issuer.pem | sudo tee /etc/nginx/conf.d/fullchain.pem
+echo "Wrote fullchain cert"
 
 # decrypt the encrypted cert from terraform
 region=$(curl -s 169.254.169.254/latest/meta-data/placement/region)
@@ -9,6 +21,7 @@ aws kms decrypt \
 --region $region  \
 --query Plaintext --output text | base64 -d  \
 | sudo tee /etc/nginx/conf.d/privkey.pem
+echo "Decrypted cert privatekey"
 
 # update nginx config to point to SSL certs
 cat <<EOF | sudo tee /etc/nginx/sites-enabled/default
@@ -27,5 +40,4 @@ server {
   }
 }
 EOF
-
-sudo service nginx reload
+echo "Configured nginx"
